chore: move ingest/lib.py to python/neuroglancer
