+++ python/neuroglancer/pipeline/precomputed.py
@@ -73,7 +73,6 @@ class Precomputed(object):
             shape=self._get_slices_shape(aligned_slices),
             dtype=self.info['data_type'])
         
-        self._storage.wait()
         offset =  self._get_offsets(aligned_slices)
         for c in self._iter_chunks(aligned_slices):
             file_path = self._chunk_to_file_path(c)
